+++ cpp/pmft/PMFTXYZ.h
@@ -32,9 +32,9 @@ public:
     */
     void accumulate(const locality::NeighborQuery* neighbor_query, 
                     quat<float>* orientations, vec3<float>* query_points,
-                    quat<float>* query_orientations, unsigned int n_query_points, 
-                    quat<float>* face_orientations, unsigned int n_faces, 
-                    const locality::NeighborList* nlist, freud::locality::QueryArgs qargs);
+                    unsigned int n_query_points, quat<float>* face_orientations,
+                    unsigned int n_faces, const locality::NeighborList* nlist,
+                    freud::locality::QueryArgs qargs);
 
     //! \internal
     //! helper function to reduce the thread specific arrays into one array