+++ ChangeLog.md
@@ -4,6 +4,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Upcoming: Changes for freud 2.0
 
+### Changed
+* Removed unused query\_orientations from PMFTXYZ.
+
 ### Fixed
 * Steinhardt uses the ThreadStorage class and properly resets memory where needed.
 * Removed all neighbor exclusion logic from Steinhardt.